files moved from integrationTest to test
