+++ src/integrationTest/java/uk/gov/hmcts/reform/judicialbooking/controller/DeleteBookingControllerIntegrationTest.java
@@ -13,7 +13,7 @@ import static org.springframework.test.web.servlet.request.MockMvcRequestBuilder
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 import static org.springframework.test.web.servlet.setup.MockMvcBuilders.standaloneSetup;
 
-public class DeleteBookingControllerIntegrationTest extends BaseTest {
+public class DeleteBookingControllerIntegrationTest extends BaseTestIntegration {
 
     private transient MockMvc mockMvc;
