+++ src/integrationTest/java/uk/gov/hmcts/reform/judicialbooking/controller/QueryBookingIntegrationTest.java
@@ -39,7 +39,7 @@ import static org.springframework.test.web.servlet.request.MockMvcRequestBuilder
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-public class QueryBookingIntegrationTest extends BaseTest {
+public class QueryBookingIntegrationTest extends BaseTestIntegration {
 
     private static final String URL = "/am/bookings/query";
