+++ src/test/java/uk/gov/hmcts/reform/judicialbooking/controller/BaseTest.java
@@ -1,0 +1,144 @@
+package uk.gov.hmcts.reform.judicialbooking.controller;
+
+import com.fasterxml.jackson.annotation.JsonInclude;
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.DeserializationFeature;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
+import com.github.tomakehurst.wiremock.WireMockServer;
+import com.launchdarkly.shaded.org.jetbrains.annotations.NotNull;
+import com.opentable.db.postgres.embedded.EmbeddedPostgres;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.context.properties.EnableConfigurationProperties;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.test.context.TestConfiguration;
+import org.springframework.boot.test.mock.mockito.MockBean;
+import org.springframework.context.ApplicationContextInitializer;
+import org.springframework.context.ApplicationListener;
+import org.springframework.context.ConfigurableApplicationContext;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.event.ContextClosedEvent;
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.MediaType;
+import org.springframework.jdbc.datasource.SingleConnectionDataSource;
+import org.springframework.lang.NonNull;
+import org.springframework.security.oauth2.client.registration.ClientRegistrationRepository;
+import org.springframework.security.oauth2.client.registration.ReactiveClientRegistrationRepository;
+import org.springframework.test.context.ActiveProfiles;
+import org.springframework.test.context.ContextConfiguration;
+import org.springframework.test.context.junit.jupiter.SpringExtension;
+import org.springframework.test.context.support.TestPropertySourceUtils;
+import uk.gov.hmcts.reform.judicialbooking.controller.utils.WiremockFixtures;
+import uk.gov.hmcts.reform.judicialbooking.util.Constants;
+
+import javax.annotation.PreDestroy;
+import javax.sql.DataSource;
+import java.io.IOException;
+import java.nio.charset.StandardCharsets;
+import java.sql.Connection;
+import java.sql.DriverManager;
+import java.sql.SQLException;
+import java.util.Properties;
+
+import static com.github.tomakehurst.wiremock.core.WireMockConfiguration.options;
+
+@ContextConfiguration(initializers = {BaseTest.WireMockServerInitializer.class})
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
+@ExtendWith(SpringExtension.class)
+//@ActiveProfiles("itest")
+@EnableConfigurationProperties
+public abstract class BaseTest {
+
+    public static final WireMockServer WIRE_MOCK_SERVER = new WireMockServer(options().dynamicPort());
+
+    protected static final ObjectMapper mapper = new ObjectMapper();
+
+    @MockBean(name = "clientRegistrationRepository")
+    private ClientRegistrationRepository getClientRegistrationRepository;
+
+    @MockBean(name = "reactiveClientRegistrationRepository")
+    private ReactiveClientRegistrationRepository getReactiveClientRegistrationRepository;
+
+    static {
+        if (!WIRE_MOCK_SERVER.isRunning()) {
+            WIRE_MOCK_SERVER.start();
+        }
+
+        mapper.registerModule(new JavaTimeModule());
+        mapper.setSerializationInclusion(JsonInclude.Include.NON_NULL);
+        mapper.configure(DeserializationFeature.ACCEPT_SINGLE_VALUE_AS_ARRAY, true);
+        // Force re-initialisation of base types for each test suite
+    }
+
+    protected static final MediaType JSON_CONTENT_TYPE = new MediaType(
+            MediaType.APPLICATION_JSON.getType(),
+            MediaType.APPLICATION_JSON.getSubtype(),
+            StandardCharsets.UTF_8
+    );
+
+    @NotNull
+    HttpHeaders getHttpHeaders() {
+        HttpHeaders headers = new HttpHeaders();
+        headers.add(Constants.CORRELATION_ID_HEADER_NAME, "38a90097-434e-47ee-8ea1-9ea2a267f51d");
+        return headers;
+    }
+
+//    @TestConfiguration
+//    static class Configuration {
+//        Connection connection;
+//
+//        @Bean
+//        public EmbeddedPostgres embeddedPostgres() throws IOException {
+//            return EmbeddedPostgres
+//                    .builder()
+//                    .start();
+//        }
+//
+//        @Bean
+//        public DataSource dataSource(@Autowired EmbeddedPostgres pg) throws Exception {
+//
+//            final Properties props = new Properties();
+//            // Instruct JDBC to accept JSON string for JSONB
+//            props.setProperty("stringtype", "unspecified");
+//            props.setProperty("user", "postgres");
+//            connection = DriverManager.getConnection(pg.getJdbcUrl("postgres"), props);
+//            return new SingleConnectionDataSource(connection, true);
+//        }
+//
+//
+//        @PreDestroy
+//        public void contextDestroyed() throws SQLException {
+//            if (connection != null) {
+//                connection.close();
+//            }
+//        }
+//    }
+
+    public static class WireMockServerInitializer
+            implements ApplicationContextInitializer<ConfigurableApplicationContext> {
+
+        private final WiremockFixtures wiremockFixtures = new WiremockFixtures();
+
+        @Override
+        public void initialize(@NonNull ConfigurableApplicationContext applicationContext) {
+
+            TestPropertySourceUtils.addInlinedPropertiesToEnvironment(
+                    applicationContext,
+                    "wiremock.server.port=" + WIRE_MOCK_SERVER.port()
+            );
+
+            try {
+                wiremockFixtures.stubIdamConfig();
+            } catch (JsonProcessingException e) {
+                throw new RuntimeException(e);
+            }
+
+            applicationContext.addApplicationListener((ApplicationListener<ContextClosedEvent>) event -> {
+                if (WIRE_MOCK_SERVER.isRunning()) {
+                    WIRE_MOCK_SERVER.shutdown();
+                }
+            });
+        }
+    }
+}
