+++ src/test/java/uk/gov/hmcts/reform/judicialbooking/util/KeyGenerator.java
@@ -1,0 +1,24 @@
+package uk.gov.hmcts.reform.judicialbooking.util;
+
+import com.nimbusds.jose.JOSEException;
+import com.nimbusds.jose.jwk.RSAKey;
+import com.nimbusds.jose.jwk.gen.RSAKeyGenerator;
+
+public final class KeyGenerator {
+
+    private static RSAKey rsaJwk;
+    private static final String KEY_ID = "23456789";
+
+    private KeyGenerator() {
+    }
+
+    public static RSAKey getRsaJwk() throws JOSEException {
+        if (rsaJwk == null) {
+            rsaJwk = new RSAKeyGenerator(2048)
+                    .keyID(KEY_ID)
+                    .generate();
+        }
+        return rsaJwk;
+    }
+
+}
